% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_normalisation.R
\name{bead_norm}
\alias{bead_norm}
\title{Cytof bead normalization via CATALYST}
\usage{
bead_norm(seu, assay = NULL, beads = "dvs", plot_res = TRUE, ...)
}
\arguments{
\item{seu}{A seurat object.}

\item{assay}{Assay from Seurat object to be used for conversion. If empty, uses DefaultAssay.}

\item{beads}{Indicate which beads were used (dvs or beta) or a custom numeric vector of bead masses (see CATALYST function).}

\item{plot_res}{Plots quality control graphs. Default: TRUE}

\item{...}{Additional parameters passed on to CATALYST::normCytof() function (see CATALYST function).}
}
\value{
Seurat Object with additional assay (beadnorm) and 2 additional meta.data columns indicating beads and bead_doublets with cells that can be removed.
}
\description{
Applies a bead normalisation on raw intensities (counts slot) of active Seurat assay using CATALYST::normCytof. A Seurat object is returned with a new assay "beadnorm" and 2 new meta.data columns for beads and potential bead/cell doublets. For details how this normalization works, please refer to https://www.bioconductor.org/packages/release/bioc/html/CATALYST.html.
}
\examples{
seu <- bead_norm(seu, beads = "dvs", k = 50) # Run bead normalisation
seu <- subset(seu, subset = bead_doublet == FALSE) # Remove beads and bead/cell doublets
}
