% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{compensate_data}
\alias{compensate_data}
\title{Compensate data based on spillover matrix save in FCS files}
\usage{
compensate_data(fcs_fs, seu, comp_matrix = NULL, fcs_matrix = 1)
}
\arguments{
\item{fcs_fs}{FlowSet containing all fcs files and compensation matrices. Must be the same FlowSet that was used for \code{\link[=create_seurat]{create_seurat()}}.}

\item{seu}{Seurat object used for \code{\link[=create_seurat]{create_seurat()}}.}

\item{comp_matrix}{Provide a compensation matrix which is the same for all samples. Names of columns and rows must match channel names indicated in panel data.frame.}

\item{fcs_matrix}{Use compensation matrices from FCS files. Can be different for every file. This parameter indicates which compensation matrix to use if there are multiple in the FCS files. Default: 1.}
}
\value{
Seurat object with new assay "comp", where compensation values were saved in slot "counts".
}
\description{
Applies compensation matrices for each channel and saves compensated values into "comp" assay in Seurat Object. Uses \code{\link[flowCore:compensation-class]{flowCore::compensate()}} function.
}
\examples{
## Option 1) Use external spillover matrix directly: same compensation for all files
# colnames must match panel data.frame; row.names are optional
comp_matrix <- read.csv("spillover_matrix.csv", row.names = 1)
seu_1 <- compensate_data(fcs_fs, seu, comp_matrix = comp_matrix)
## Option 2) Use spillover matrix saved in FCS files: same compensation for all files
comp_matrix <- spillover(fcs_fs[[1]])[[3]] # Matrix 3 from first FCS file
seu_2 <- compensate_data(fcs_fs, seu, comp_matrix = comp_matrix)
## Option 3) Use spillover matrix saved in FCS files: compensation matrix used from individual FCS files, thus can be different for each file
# Check different matrices using:
names(spillover(fcs_fs[[1]]))
# Pick the correct spillover matrix for compensation and provide with fcs_matrix parameter
seu <- compensate_data(fcs_fs, seu, fcs_matrix = 3)
}
