% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_seurat}
\alias{convert_seurat}
\title{Converts a Seurat object to a FlowFrame, FlowSet or SingleCellExperiment Object.}
\usage{
convert_seurat(
  seu,
  to = c("FF", "FS", "SCE"),
  assay = NULL,
  slot = "data",
  split_by
)
}
\arguments{
\item{seu}{A seurat object.}

\item{to}{Can be either FF for FlowFrame, FS for FlowSet, or SCE for SingleCellExperiment.}

\item{assay}{Assay from Seurat object to be used for conversion. If empty, uses DefaultAssay.}

\item{slot}{Slot from Seurat object to be used for conversion. If empty, uses "data" slot.}

\item{split_by}{Indicate which seu@meta.data column to use to split data into individual FlowFrames to be used in a FlowSet.}
}
\value{
Either a FlowFrame, FlowSet, or SingleCellExperiment Object.
}
\description{
Takes Seurat object as input and returns FlowFrame, FlowSet or SingleCellExperiment Object. These can be used for integration of third-party tools. By default uses DefaultAssay and Slot from Seurat object as Matrix. Only channels used in matrix and present in panel are used for conversion. Unused Channels will not be included.
}
\examples{
sce = convert_seurat(seu, assay = "comp", to = "SCE")
}
