% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{create_flowset}
\alias{create_flowset}
\title{Load multiple FCS files as flowSet Object}
\usage{
create_flowset(fcs_dir, metadata)
}
\arguments{
\item{fcs_dir}{Directory containing all FCS files.}

\item{metadata}{Data.frame with metadata for each FCS file (rows). Must contain columns "sample_id" and "file_name".}
}
\value{
A flowCore FlowSet object.
}
\description{
Uses \code{\link[flowCore:read.flowSet]{flowCore::read.flowSet()}} to generate a flowSet based on files within a certain directory.
In addition, this function will give each cell a unique name based on the sample_id so it can be identified later in downstream analyses.
This flowSet Object should be stored in order to export FCS files based on unique cell IDs later.
}
\examples{
create_flowset("fcs_files_directory", metadata)
}
