% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{create_seurat}
\alias{create_seurat}
\title{Converts a flowCore flowSet containing multiple flowFrames into a Seurat object}
\usage{
create_seurat(fcs_fs, panel, metadata = NULL, ...)
}
\arguments{
\item{fcs_fs}{A flowCore flowSet generated with create_flowset.}

\item{panel}{A panel containing all information for the markers. See vignette for more information on how to generate a panel_df. Function will only use channels indicated in panel$fcs_colnames and will rename channels to panel$antigen.}

\item{metadata}{Data.frame with metadata for each FCS file (rows). Must contain columns "sample_id" and "file_name" and each row must correspond to the flowFrames in the flowSet.}

\item{...}{Additional parameters that will be passed on to Seurat::CreateSeuratObject.}
}
\value{
A Seurat Object of all merged FCS files.
}
\description{
Uses sample metadata to generate cell-level metadata. Each row = one FCS file. Must contain "file_name" and "sample_id" columns. Uses panel data.frame to subset a flowSet to channels that should be analyzed, and channels that will not be used for downstream analysis. It's important that panel$fcs_colnames contains all channels that should be used. For additional explanation of panel_df, see vignette.
}
\details{
Channel names can be checked in the flowFrame using: colnames(fcs_fs@frames$fcs_id@exprs). Raw FCS data is stored in seu@assays$fcs. Raw FCS data of unused channels is stored in seu@assays$unused. Panel data is stored in seu@misc slot.
}
\examples{
seu = create_seurat(fcs_fs, panel, metadata)
}
