% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_pseudobulk.R
\name{DE_pseudobulk}
\alias{DE_pseudobulk}
\title{Computes differentially expressed proteins based on median fluorescence intensity per group using limma}
\usage{
DE_pseudobulk(seu, group_by = "sample_id", fixed_vars, contrast, min.cells = 3)
}
\arguments{
\item{seu}{Seurat object.}

\item{group_by}{Indicate group which is used for pseudobulking using \code{\link[=median_expression]{median_expression()}}. Default: sample_id.}

\item{fixed_vars}{Fixed variables for the linear model.}

\item{contrast}{Indicate which contrast to make, also see limma vignette. For example, attribute1condition1-attribute1condition2. Second mention is always the reference.}

\item{min.cells}{Minimum number of cells that should be in a group. If less cells, the group is dropped.}
}
\value{
Data.frame containing \code{\link[limma:toptable]{limma::topTable()}} results.
}
\description{
This function calculates the median fluorescence intensity per group using \code{\link[=median_expression]{median_expression()}} to create a "pseudobulk" matrix. To detect differentially expressed proteins, limma is used: \code{\link[limma:lmFit]{limma::lmFit()}} to fit the linear model and \code{\link[limma:contrasts.fit]{limma::contrasts.fit()}}, \code{\link[limma:ebayes]{limma::eBayes()}} to make contrasts. Returns data.frame of \code{\link[limma:toptable]{limma::topTable()}} results.
}
\examples{
# compute differential expression
DE_res <- DE_pseudobulk(seu, fixed_vars = "condition", contrast = "conditionA-conditionB")
}
