% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_detection.R
\name{detect_aggregate_channels}
\alias{detect_aggregate_channels}
\title{Detect channel combinations that potentially contain aggregates}
\usage{
detect_aggregate_channels(
  seu,
  threshold = 0.7,
  pos_threshold = 1,
  percent_cells = 0.005
)
}
\arguments{
\item{seu}{Seurat object.}

\item{threshold}{Threshold for Pearson R. Channel combinations with a Pearson R higher than this threshold will be tagged potentially containing aggregates.}

\item{pos_threshold}{Treshold that distinguishes positive and negative events. Default: 1.}

\item{percent_cells}{Only tag channels as potentially containing aggregates if at least a certain percentage of all cells are double positive. This is too avoid high Pearson R in presence of very limited number of cells. Default: 0.5\% (= 0.005)}
}
\value{
List of correlation matrix with pearson R, correlation matrix with 1 for above threshold and 0 for below and data.frame with channel combinations that may contain aggregates.
}
\description{
Channels will be subset to positive events only, since aggregates can only be detected in positive events. Pearson correlation is performed for every channel combination and a threshold is applied to determine which channel combinations may contain aggregates.
}
\examples{
# identify channel combinations that potentially contain aggregates
problem_channels <- detect_aggregate_channels(seu)
# pheatmap can be used to quickly plot correlation matrices
library(pheatmap)
pheatmap(problem_channels[[1]])
pheatmap(problem_channels[[2]])
}
