% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_detection.R
\name{detect_aggregates}
\alias{detect_aggregates}
\title{Labels potential aggregates based on modified RANSAC model in Seurat object (seu$aggregate)}
\usage{
detect_aggregates(
  seu,
  aggregate_channels,
  score_threshold = 2,
  fit_threshold = 0.25,
  pos_threshold = 1,
  ...
)
}
\arguments{
\item{seu}{Seurat object.}

\item{aggregate_channels}{A data.frame with columns giving the channel combinations that potentially contain aggregates. This can be determine with \code{\link[=detect_aggregate_channels]{detect_aggregate_channels()}} which returns a list. Element 3 of the list is this data.frame (aggregate_channels[\link{3}]).}

\item{score_threshold}{Number of channel combinations in which a cells should at be labeled as an aggregate (aggregate_score) to regard it as an aggregate. The higher the number, the less likely false positives will occur, but the more likely aggregates could be missed. Default: >=2.}

\item{fit_threshold}{Threshold for residuals (y - predicted) to be counted as aggregates in modified RANSAC model. Also see \code{\link[=run_modified_ransac]{run_modified_ransac()}}.}

\item{pos_threshold}{Treshold that distinguishes positive and negative events. Should be the value as provided for \code{\link[=detect_aggregate_channels]{detect_aggregate_channels()}}. Default: 1.}

\item{...}{Arguments that will be passed on to \code{\link[=run_modified_ransac]{run_modified_ransac()}}.}
}
\value{
Seurat object with new cell-level meta.data "aggregate_score" and "aggregate".
}
\description{
Problematic channels that may contain aggregates can be identified with \code{\link[=detect_aggregate_channels]{detect_aggregate_channels()}} function. This function runs a modified RANSAC algorithm (see \code{\link[=run_modified_ransac]{run_modified_ransac()}}) on data of each channel combinations that may contain aggregates and scores the probability of events being aggregates. If cells are labelled as aggregates in several channels (score_threshold), cells are regarded as aggregates, which is stored in Seurat Object (seu$aggregate). The number of channels in which a cells is labelled as aggregate is stored in seu$aggregate_score.
}
\examples{
# detect aggregates in problematic channels
seu <- detect_aggregates(seu, problem_channels[[3]])
# remove aggregate events from Seurat object
seu <- subset(seu, subset = aggregate == FALSE)
}
