% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Labels potentially noisy or outlier events in Seurat object (seu$outlier)}
\usage{
detect_outliers(
  seu,
  score_threshold = 0.7,
  ndim = NULL,
  ntrees = 100,
  seed = 42,
  ...
)
}
\arguments{
\item{seu}{Seurat object.}

\item{score_threshold}{The threshold by which an event is called outlier based on isolation forest score. Usually between 0.6 - 0.8. Default: 0.7.}

\item{ndim}{Isolation tree number of dim. Also see ?isolation.forest. Default: NULL.}

\item{ntrees}{Isolation tree number of trees. Also see ?isolation.forest. Default: 100.}

\item{seed}{Set seed for reproducible results. Default: 42.}

\item{...}{Additional parameters passed on to isotree::isolation.forest function.}
}
\value{
Seurat object with new cell-level meta.data "outlier" and "outlier_score".
}
\description{
Uses isolation forest to score events based on their probability to be an outlier. Scores and predicted outliers are stored in Seurat meta.data "outlier" and "outlier_score".
}
\examples{
# score and predict outliers based on threshold
seu <- detect_outliers(seu, threshold = 0.6)
# remove outliers from Seurat object
seu <- subset(seu, subset = outlier == FALSE)
}
