% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{differential_abundance}
\alias{differential_abundance}
\title{Computes differential abundances between conditions using edgeR}
\usage{
differential_abundance(
  seu,
  attribute = "seurat_clusters",
  group_by = "sample_id",
  formula,
  contrast,
  check_coeff = FALSE
)
}
\arguments{
\item{attribute}{Attribute used to calculate differential abundance. Must be present in meta.data of Seurat Object. Default: seurat_clusters.}

\item{group_by}{Attribute used to group cells. Default: sample_id.}

\item{formula}{Formula for GLM, e.g. "~condition1+condition2+condition3". Conditions must be present as attritubes in seu@meta.data.}

\item{contrast}{Contrast for coefficients, e.g. c(0,1,-1). To determine how to set the desired contrast, use check_coeff = TRUE.}

\item{check_coeff}{Return coefficients to pick contrasts.}

\item{cellcounts}{Matrix with cellcounts per attribute.}
}
\value{
Data.frame containing topTable results of contrast, corrected for multiple comparison using BH.
}
\description{
This function will use \code{\link[edgeR:glmfit]{edgeR::glmFit()}} to model number of cells per group given attribute/feature x condition (e.g. seurat_clusters x treatment) and \code{\link[edgeR:glmfit]{edgeR::glmLRT()}} to make contrast between all conditions. Returns the topTags results corrected for multiple comparison using BH and sorted by pvalue.
}
\examples{
' # first check the coefficiencs
differential_abundance(seu, attribute = "seurat_clusters", group_by = "sample_id", formula = as.formula("~0+condition"), check_coef = TRUE)
'# calculate differential abundance
res <- differential_abundance(seu, attribute = "seurat_clusters", group_by = "sample_id", formula = as.formula("~0+condition"), contrast = c(1, -1))
res
}
