% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{export_fcs}
\alias{export_fcs}
\title{Exports FCS files from Seurat object.}
\usage{
export_fcs(seu, filename, assay = NULL, slot = "data")
}
\arguments{
\item{seu}{A seurat object.}

\item{filename}{Path and filename where FCS file should be written to.}

\item{assay}{Assay from Seurat object to be used for conversion. If empty, uses DefaultAssay.}

\item{slot}{Slot from Seurat object to be used for conversion. If empty, uses "data" slot.}
}
\description{
This function uses a specified assay and slot from a Seurat object, converts the Object to a flowFrame, and exports the flowFrame as a FCS file. This can be used, for example, to visualize specific clusters or cells in a traditional Flow Cytometry Software.
}
\examples{
# subset Seurat object to specific cluster
seu_sub <- subset(seu, subset = seurat_clusters == 1)
# export FCS file containing cells from this cluster using raw data
export_fcs(seu_sub, filename = "cluster_1.fcs", assay = "fcs", slot = "counts")
}
