% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{median_expression}
\alias{median_expression}
\title{Calculates median fluorescence intensity per group based on "data" slot in DefaultAssay(seu)}
\usage{
median_expression(seu, group_by = "ident")
}
\arguments{
\item{seu}{Seurat object.}

\item{group_by}{Indicate a column in meta.data of Seurat Object by which the data should be grouped. Default: Idents(Seu)}
}
\value{
Matrix with median expression values across different groups.
}
\description{
Calculates median fluorescence intensity per group based on "data" slot in DefaultAssay(seu)
}
\examples{
sample_mfi <- median_expression(seu, group_by = "sample_id")
}
