% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cellnumber}
\alias{plot_cellnumber}
\title{Barplot cell numbers}
\usage{
plot_cellnumber(seu, group_by = "sample_id", return_table = FALSE)
}
\arguments{
\item{seu}{Seurat object}

\item{return_table}{Return table with frequencies instead of ggplot2 plot. Default: FALSE.}

\item{group}{Group to plot number of cells. Default: sample_id.}
}
\value{
ggplot2 plot, or table if return_table = TRUE
}
\description{
Barplot cell numbers
}
\examples{
plot_cellnumber(seu)
}
