% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cyto}
\alias{plot_cyto}
\title{Plot Cytometry-Style plots on Seurat objects}
\usage{
plot_cyto(
  seu,
  x,
  y,
  style = "2d_density",
  slot = "data",
  assay = NULL,
  scale = "linear",
  limits = NULL,
  color = NULL,
  pt_size = 0.1,
  rasterize = FALSE,
  alpha = 1,
  bins = 200,
  biexp_pos = 4.5,
  biexp_neg = 0,
  biexp_widthBasis = -10
)
}
\arguments{
\item{seu}{Seurat object.}

\item{x}{Which marker to plot on x axis.}

\item{y}{Which marker to plot on y axis (only required for style = 2d_density or points)}

\item{style}{Choose from: "2d_density", "point", or "density". Default: 2d_density.}

\item{slot}{Choose which Seurat slot to use. Default: data (transformed).}

\item{assay}{Choose which Seurat assay to use. Default: DefaultAssay(seu).}

\item{scale}{Choose which scale to use for x and y axis (linear, log, biexp). Default: linear.}

\item{limits}{List of x and y limits. Example: list(x = c(0,5), y = c(0,5)).}

\item{color}{Color cells based on meta.data column in Seurat object, e.g. "sample_id".}

\item{pt_size}{When using style = "point", this parameter adjusts the point size (Default: 0.2).}

\item{rasterize}{Rasterize plot. Useful for plotting large datasets (Default: FALSE).}

\item{alpha}{When using style = "point", this parameter adjusts the alpha (Default: 1).}

\item{bins}{When using style = "2d_density", this parameter adjusts the number of bins.}

\item{biexp_pos}{If scale = "biexp", indicate biexp_pos here (see \code{\link[=transform_biexp]{transform_biexp()}}).}

\item{biexp_neg}{If scale = "biexp", indicate biexp_neg here (see \code{\link[=transform_biexp]{transform_biexp()}}).}

\item{biexp_widthBasis}{If scale = "biexp", indicate biexp_widthBasis here (see \code{\link[=transform_biexp]{transform_biexp()}}).}
}
\value{
ggplot2 plot.
}
\description{
Plot Cytometry-Style plots on Seurat objects
}
\examples{
plot_cyto(seu, x = "CD3", y = "CD4")
plot_cyto(seu, x = "CD3", y = "CD4", style = "point", color = "sample_id")
plot_cyto(seu, x = "CD3", style = "density", color = "sample_id")
}
