% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Stacked barplot with percentage or number of cells per attribute by attribute}
\usage{
plot_frequency(
  seu,
  attribute_1,
  attribute_2,
  n_cells = FALSE,
  return_table = FALSE
)
}
\arguments{
\item{seu}{Seurat object.}

\item{attribute_1}{Attribute used for y axis.}

\item{attribute_2}{Attribute used for x axis.}

\item{n_cells}{Plot number of cells instead of frequencies. Default: FALSE.}

\item{return_table}{Return table with frequencies instead of ggplot2 plot. Default: FALSE.}
}
\value{
ggplot2 plot, or table if return_table = TRUE
}
\description{
Stacked barplot with percentage or number of cells per attribute by attribute
}
\examples{
plot_frequency(seu, "seurat_clusters", "sample_id")
}
