% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{remove_outliers_manual}
\alias{remove_outliers_manual}
\title{Manually remove noisy events using positive and/or negative thresholds}
\usage{
remove_outliers_manual(seu, thresholds, negative = TRUE)
}
\arguments{
\item{seu}{Seurat object.}

\item{thresholds}{Named vector of thresholds per marker.}

\item{negative}{Negative or positive thresholds, i.e. remove events below or above the indicated value. Default: TRUE (= remove events below values).}
}
\value{
Seurat object with cells removed that are below or above threshold.
}
\description{
Remove cells below or above certain threshold for each channel. These can be artefacts with very bright or dim fluorescence.
}
\examples{
# Set positive and negative threshold for each channel
thresholds_neg = c("CD3" = -3, "CD4" = -2)
thresholds_pos = c("CD3" = 4, "CD4" = 2)
# Remove events below indicated thresholds
seu_manual = remove_outliers_manual(seu, thresholds)
# Remove events above indicated thresholds
seu_manual = remove_outliers_manual(seu_manual, thresholds, negative = FALSE)
}
