% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSOM.R
\name{run_FlowSOM}
\alias{run_FlowSOM}
\title{Computes FlowSOM- and metaclustering}
\usage{
run_FlowSOM(
  seu,
  assay = NULL,
  slot = "data",
  features = NULL,
  metaclusters = 10,
  xdim = 10,
  ydim = 10,
  ...
)
}
\arguments{
\item{seu}{A seurat object.}

\item{assay}{Assay from Seurat object to be used for conversion. If empty, uses DefaultAssay.}

\item{slot}{Slot from Seurat object to be used for conversion. If empty, uses "data" slot.}

\item{features}{Indicate which features should be used for FlowSOM clustering. Ideally, the features are identical to the features used for dimensionality reduction. Default: NULL, which means that all features are used.}

\item{metaclusters}{The number of metaclusters to be computed. Default: 10.}

\item{xdim}{The X dimension of the self-organizing map. Default: 10.}

\item{ydim}{The Y dimension of the self organizing map. Default: 10.}

\item{...}{Additional parameters passed on to FlowSOM::BuildSOM() function.}
}
\value{
Seurat Object containing FlowSOM- and metaclusters in seu@meta.data.
}
\description{
This function computes FlowSOM- and metaclustering using the FlowSOM packages and writes SOM cluster and metacluster information into meta.data of Seurat object. Data should be compensated and transformed.
}
\examples{
seu <- run_FlowSOM(seu)
}
