% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_detection.R
\name{run_modified_ransac}
\alias{run_modified_ransac}
\title{Modified Random Sample Consensus (RANSAC) to model aggregate events in channels}
\usage{
run_modified_ransac(
  formula,
  data,
  min_to_fit = 10,
  max_iteration = 10,
  fit_threshold = 0.5,
  data_close = 10,
  seed = 42
)
}
\arguments{
\item{formula}{Formula for linear model containing dependent and independent variable.}

\item{data}{Data to run model on (events that pass the pos_threshold).}

\item{min_to_fit}{Minimum number of events to sample and fit the linear models. Default: 10.}

\item{fit_threshold}{Threshold for residuals (y - predicted) to be counted as inliers.}

\item{data_close}{Minimum data points counted as inliers to consider the model.}

\item{seed}{Set seed for reproducible results. Default: 42.}

\item{max_iterations}{Number of iterations RANSAC should run.}
}
\value{
Best fitting model for formula and data.
}
\description{
Channels will be subset to positive events only, since aggregates can only be detected in positive events. A modified RANSAC is run that samples on events in the upper top right quadrant (events > 50\% of fluorescence in both channels) and only allows linear models with a slope between 0.75 - 1.25 since aggregates follow an almost x = y model. Function returns the best fitting model.
}
