% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_arcsinh}
\alias{transform_arcsinh}
\title{Arcsinh transformation using cofactors for each individual channel}
\usage{
transform_arcsinh(matrix, panel)
}
\arguments{
\item{matrix}{Provide matrix of flow data. E.g. "Seurat::GetAssayData(seu)".}

\item{panel}{Provide data.frame with panel and transformation information. See vignette for details how to generate a panel_df.}
}
\value{
Transformed matrix that can be added to Seurat object.
}
\description{
Transforms values for every channel based on cofactor provided in panel data.frame (arcsinh_cofactor). If column not present, uses default cofactor: 5. Function used: y = arcsinh(x/cofactor). It's best to use the wrapper function \code{\link[=transform_data]{transform_data()}} instead of using this function directly.
}
\examples{
seu <- SetAssayData(seu, slot = "data", new.data = transform_arcsinh(matrix, panel))
}
