% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_biexp}
\alias{transform_biexp}
\title{Biexponential transformation based on flowWorkspace::flowjo_biexp}
\usage{
transform_biexp(matrix, panel)
}
\arguments{
\item{matrix}{Provide matrix of flow data. E.g. "Seurat::GetAssayData(seu)".}

\item{panel}{Provide data.frame with panel and transformation information. See vignette for details how to generate a panel_df.}
}
\value{
Transformed matrix that can be added to Seurat object.
}
\description{
Transforms values for every channel based on panel data frame. Biexp can be used with 3 parameters (columns) per channel based on FlowJo 10: biexp_pos, biexp_neg, biexp_width. If a column is not provided, function will use default values. See \code{\link[flowWorkspace:flowjo_biexp]{flowWorkspace::flowjo_biexp()}} function for default values. It's best to use the wrapper function \code{\link[=transform_data]{transform_data()}} instead of using this function directly.
}
\examples{
seu <- SetAssayData(seu, slot = "data", new.data = transform_biexp(matrix, panel))
}
