% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{compensate_data}
\alias{compensate_data}
\title{Compensate data based on spillover matrix save in FCS files}
\usage{
compensate_data(fcs_fs, seu, comp_matrix = 1)
}
\arguments{
\item{fcs_fs}{FlowSet containing all fcs files and compensation matrices. Must be the same FlowSet that was used for \code{\link[=create_seurat]{create_seurat()}}.}

\item{seu}{Seurat object used for \code{\link[=create_seurat]{create_seurat()}}.}

\item{comp_matrix}{If multiple compensation matrices are present in fcs file, indicate which one to use (check with names(spillover(fcs_fs[\link{x}]))).}
}
\value{
Seurat object with new assay "comp", where compensation values were saved in slot "counts".
}
\description{
Applies compensation matrices for each channel and saves compensated values into "comp" assay in Seurat Object. Uses \code{\link[flowCore:compensation-class]{flowCore::compensate()}} function.
}
\examples{
# Check if there are multiple spillover matrices saved in FCS files
names(spillover(fcs_fs[[1]]))
# Pick the correct spillover matrix for compensation
seu = compensate_data(fcs_fs, seu, comp_matrix = 3)
}
